Extract types from server/types
