+++ platform/config/index.ts
@@ -1,3 +1,4 @@
 export { ConfigService } from './ConfigService';
 export { RawConfigService } from './RawConfigService';
 export { Env } from './Env';
+export { ConfigWithSchema } from './ConfigWithSchema';
