+++ platform/types.ts
@@ -1,5 +1,3 @@
-import { Observable } from 'rxjs';
-
 /**
  * TypeScript's type system is structural, but for some use-cases we want to be
  * able to create "branded types", e.g. create two types `Id` and `PluginName`
@@ -61,17 +59,11 @@ export declare class As<S extends string> {
 
 // TODO inline all of these
 import * as schemaLib from './lib/schema';
-import { ConfigService, Env } from './config';
-import { Router, RouterOptions, HttpModule } from './server/http';
-import { KibanaConfig, KibanaModule } from './server/kibana';
-import {
-  ElasticsearchService,
-  ElasticsearchConfigs,
-  ElasticsearchModule
-} from './server/elasticsearch';
+import { ConfigService } from './config';
+import { HttpModule } from './server/http';
+import { KibanaModule } from './server/kibana';
+import { ElasticsearchModule } from './server/elasticsearch';
 import { LoggerFactory } from './logger';
-
-export type ElasticsearchClusterType = 'data' | 'admin';
 
 export type Schema = typeof schemaLib;
 
@@ -94,95 +86,6 @@ export interface KibanaCoreModules {
   http: HttpModule;
   configService: ConfigService;
   logger: LoggerFactory;
-}
-
-export interface KibanaPluginFeatures {
-  /**
-   * Plugin-scoped logger
-   */
-  logger: LoggerFactory;
-
-  /**
-   * Core Kibana utilities
-   */
-  util: {
-    schema: Schema;
-  };
-
-  /**
-   * Core Elasticsearch functionality
-   */
-  elasticsearch: {
-    service: ElasticsearchService;
-    config$: Observable<ElasticsearchConfigs>;
-  };
-  kibana: {
-    config$: Observable<KibanaConfig>;
-  };
-
-  /**
-   * Core HTTP functionality
-   */
-  http: {
-    /**
-     * Create and register a router at the specified path.
-     *
-     * The return value of the `onRequest` router option will be injected as the
-     * first param in any route handler registered on the router.
-     */
-    createAndRegisterRouter: <T>(
-      path: string,
-      options: RouterOptions<T>
-    ) => Router<T>;
-  };
-
-  /**
-   * Core configuration functionality, enables fetching a subset of the config.
-   */
-  config: {
-    /**
-     * Reads the subset of the config at the specified `path` and validates it
-     * against the schema created by calling the static `createSchema` on the
-     * specified `ConfigClass`.
-     *
-     * @param path The path to the desired subset of the config.
-     * @param ConfigClass A class (not an instance of a class) that contains a
-     * static `createSchema` that will be called to create a schema that we
-     * validate the config at the given `path` against.
-     */
-    create: <Schema extends schemaLib.Any, Config>(
-      ConfigClass: ConfigWithSchema<Schema, Config>
-    ) => Observable<Config>;
-    createIfExists: <Schema extends schemaLib.Any, Config>(
-      ConfigClass: ConfigWithSchema<Schema, Config>
-    ) => Observable<Config | undefined>;
-  };
-}
-
-/**
- * Interface that defines the static side of a config class.
- *
- * (Remember that a class has two types: the type of the static side and the
- * type of the instance side, see https://www.typescriptlang.org/docs/handbook/interfaces.html#difference-between-the-static-and-instance-sides-of-classes)
- *
- * This can't be used to define the config class because of how interfaces work
- * in TypeScript, but it can be used to ensure we have a config class that
- * matches whenever it's used.
- */
-export interface ConfigWithSchema<S extends schemaLib.Any, Config> {
-  /**
-   * Any config class must define a schema that validates the config, based on
-   * the injected `schema` helper.
-   */
-  createSchema: (schema: Schema) => S;
-
-  /**
-   * @param validatedConfig The result from calling the static `createSchema`
-   * above. This config is validated before the config class is instantiated.
-   * @param env An instance of the `Env` class that defines environment specific
-   * variables.
-   */
-  new (validatedConfig: schemaLib.TypeOf<S>, env: Env): Config;
 }
 
 export interface CoreService {
